chore: reorganize unit tests for oauth-v5 package
