+++ packages/oauth-v5/test/unit/authorizations/info.unit.test.js
@@ -4,7 +4,7 @@
 const cli = require('heroku-cli-util')
 const nock = require('nock')
 const {expect} = require('chai')
-const cmd = require('../../lib/commands/authorizations/info')
+const cmd = require('../../../lib/commands/authorizations/info')
 
 const distanceInWordsToNow = require('date-fns/distance_in_words_to_now')
