+++ packages/oauth-v5/test/unit/sessions/index.unit.test.js
@@ -1,10 +1,10 @@
+/* eslint-env mocha */
 'use strict'
-/* globals beforeEach afterEach */
 
 const cli = require('heroku-cli-util')
 const {expect} = require('chai')
 const nock = require('nock')
-const cmd = require('../../lib/commands/clients/index')
+const cmd = require('../../../lib/commands/sessions')
 
 describe('clients', () => {
   let api
@@ -20,31 +20,31 @@ describe('clients', () => {
   describe('with authorizations', () => {
     beforeEach(() => {
       api = api
-        .get('/oauth/clients')
-        .reply(200, [{name: 'awesome', id: 'f6e8d969-129f-42d2-854b-c2eca9d5a42e', redirect_uri: 'https://myapp.com'}])
+        .get('/oauth/sessions')
+        .reply(200, [{description: 'Session @ 166.176.184.223', id: 'f6e8d969-129f-42d2-854b-c2eca9d5a42e'}])
     })
 
-    it('lists the clients', () => {
+    it('lists the sessions', () => {
       return cmd.run({flags: {}})
-        .then(() => expect(cli.stdout).to.equal('awesome  f6e8d969-129f-42d2-854b-c2eca9d5a42e  https://myapp.com\n'))
+        .then(() => expect(cli.stdout).to.equal('Session @ 166.176.184.223  f6e8d969-129f-42d2-854b-c2eca9d5a42e\n'))
     })
 
-    it('lists the clients as json', () => {
+    it('lists the sessions as json', () => {
       return cmd.run({flags: {json: true}})
-        .then(() => expect(JSON.parse(cli.stdout)[0], 'to satisfy', {name: 'awesome'}))
+        .then(() => expect(JSON.parse(cli.stdout)[0], 'to satisfy', {description: 'Session @ 166.176.184.223'}))
     })
   })
 
   describe('without authorizations', () => {
     beforeEach(() => {
       api = api
-        .get('/oauth/clients')
+        .get('/oauth/sessions')
         .reply(200, [])
     })
 
     it('shows no clients message', () => {
       return cmd.run({flags: {}})
-        .then(() => expect(cli.stdout).to.equal('No OAuth clients.\n'))
+        .then(() => expect(cli.stdout).to.equal('No OAuth sessions.\n'))
     })
   })
 })
